# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Testing Shop API",
    "version": "1.0.0",
    "description": "Demo e-shop backend with fault injection for API/UI/Unit testing scenarios."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://{apiHost}",
      "description": "Deployment host",
      "variables": {
        "apiHost": {
          "default": "example.azurecontainerapps.io"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Money": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "currencyCode": {
            "type": "string",
            "example": "CZK"
          }
        },
        "required": [
          "amount",
          "currencyCode"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "inStock": {
            "type": "integer"
          },
          "active": {
            "type": "boolean"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "inStock",
          "active",
          "price"
        ]
      },
      "StorefrontCatalogResponse": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "priceBounds": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "currencyCode": {
                "type": "string"
              }
            },
            "required": [
              "min",
              "max",
              "currencyCode"
            ]
          }
        },
        "required": [
          "products",
          "priceBounds"
        ]
      },
...
      "AdminFault": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "UI",
              "API",
              "Unit"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "latencyMs": {
            "type": "number",
            "nullable": true
          },
          "failureRate": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "key",
          "name",
          "description",
          "level",
          "enabled",
          "latencyMs",
          "failureRate"
        ]
      }
    }
  },
...
  "paths": {
    "/admin/faults": {
      "get": {
        "tags": [
          "Faults"
        ],
        "summary": "List all fault configurations for admin/tester.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fault config list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminFault"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/faults/{key}": {
      "patch": {
        "tags": [
          "Faults"
        ],
        "summary": "Update a single fault config by key.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "latencyMs": {
                    "type": "number",
                    "nullable": true
                  },
                  "failureRate": {
                    "type": "number",
                    "nullable": true
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "level": {
                    "type": "string",
                    "enum": [
                      "UI",
                      "API",
                      "Unit"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated fault config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFault"
                }
              }
            }
          },
          "400": {
            "description": "Invalid key or payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List all products for admin editing.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product list for admin.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create product as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "inStock": {
                    "type": "integer"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "price": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "number"
                      },
                      "currencyCode": {
                        "type": "string",
                        "example": "CZK"
                      }
                    },
                    "required": [
                      "amount",
                      "currencyCode"
                    ]
                  }
                },
                "required": [
                  "name",
                  "description",
                  "inStock",
                  "active",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product by id as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "inStock": {
                    "type": "integer"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "price": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "number"
                      },
                      "currencyCode": {
                        "type": "string",
                        "example": "CZK"
                      }
                    },
                    "required": [
                      "amount",
                      "currencyCode"
                    ]
                  }
                },
                "required": [
                  "name",
                  "description",
                  "inStock",
                  "active",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product by id as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted."
          },
          "400": {
            "description": "Invalid id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get storefront product catalog with filtering/sorting/localization.",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search term for name/description."
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            },
            "description": "Storefront language."
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "name-asc",
                "name-desc",
                "price-asc",
                "price-desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "priceMin",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "priceMax",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Storefront catalog response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}